[refactoring/daengle-26] User 견적서 관련 요청-응답 DTO 도메인 모듈로부터 가져옴과 동시에 Mapper 생성
