+++ daengle-user-api/src/main/java/ddog/user/application/AccountService.java
@@ -7,14 +7,14 @@ import ddog.domain.pet.Pet;
 import ddog.domain.user.User;
 import ddog.user.application.mapper.PetMapper;
 import ddog.user.application.mapper.UserMapper;
-import ddog.user.presentation.dto.response.BreedInfo;
-import ddog.user.presentation.dto.response.PetInfo;
-import ddog.user.presentation.dto.response.UserProfileInfo;
+import ddog.user.presentation.account.dto.*;
+import ddog.user.presentation.account.dto.BreedInfo;
+import ddog.user.presentation.account.dto.PetInfo;
+import ddog.user.presentation.account.dto.UserProfileInfo;
 import ddog.persistence.port.AccountPersist;
 import ddog.persistence.port.PetPersist;
 import ddog.persistence.port.UserPersist;
-import ddog.user.presentation.dto.response.ValidResponse;
-import ddog.user.presentation.dto.request.*;
+import ddog.user.presentation.account.dto.ValidResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
