+++ daengle-user-api/src/main/java/ddog/user/presentation/account/dto/DeletePetId.java
@@ -1,4 +1,4 @@
-package ddog.user.presentation.dto.request;
+package ddog.user.presentation.account.dto;
 
 import lombok.Getter;
