+++ daengle-user-api/src/main/java/ddog/user/presentation/account/dto/ModifyPetInfo.java
@@ -1,4 +1,4 @@
-package ddog.user.presentation.dto.request;
+package ddog.user.presentation.account.dto;
 
 import ddog.domain.pet.Breed;
 import ddog.domain.pet.Gender;
@@ -6,8 +6,9 @@ import ddog.domain.pet.Weight;
 import lombok.Getter;
 
 @Getter
-public class AddPetInfo {
+public class ModifyPetInfo {
 
+    private Long id;
     private String image;
     private String name;
     private int birth;
