+++ daengle-user-api/src/main/java/ddog/user/presentation/account/dto/PetInfo.java
@@ -1,4 +1,4 @@
-package ddog.user.presentation.dto.response;
+package ddog.user.presentation.account.dto;
 
 import ddog.domain.pet.Breed;
 import ddog.domain.pet.Gender;
