+++ daengle-user-api/src/main/java/ddog/user/presentation/estimate/EstimateControllerResp.java
@@ -1,11 +1,11 @@
-package ddog.domain.estimate.enums;
+package ddog.user.presentation.estimate;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public enum UserEstimateControllerResp {
+public enum EstimateControllerResp {
     GENERAL_GROOMING_REGISTRATION("일반 사용자 미용 견적서 등록 완료"),
     DESIGNATION_GROOMING_REGISTRATION("지정 미용사 사용자 미용 견적서 등록 완료"),
     GENERAL_CARE_REGISTRATION("일반 사용자 진료 견적서 등록 완료"),
