+++ daengle-user-api/src/main/java/ddog/user/presentation/estimate/dto/GeneralGroomingEstimateReq.java
@@ -1,4 +1,4 @@
-package ddog.domain.estimate.dto.request;
+package ddog.user.presentation.estimate.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 
@@ -8,7 +8,7 @@ import lombok.Getter;
 import java.time.LocalDateTime;
 
 @Getter
-public class UserGeneralCareEstimateReq {
+public class GeneralGroomingEstimateReq {
 
     private String address;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss", timezone = "Asia/Seoul")
@@ -19,7 +19,8 @@ public class UserGeneralCareEstimateReq {
     private String significant;
     private int birth;
     private Weight weight;
-    private String symptoms;
+
+    private String desiredStyle;
     private String requirements;
     private String userImage;
     private String nickname;
